move OrderingUtil usage to base class
