+++ src/AbstractSubscription.js
@@ -1,9 +1,12 @@
 import EventEmitter from 'eventemitter3'
 import debugFactory from 'debug'
+import { Errors, Utils } from 'streamr-client-protocol'
 
 import VerificationFailedError from './errors/VerificationFailedError'
 import InvalidSignatureError from './errors/InvalidSignatureError'
+import EncryptionUtil from './EncryptionUtil'
 
+const { OrderingUtil } = Utils
 const debug = debugFactory('StreamrClient::AbstractSubscription')
 
 let subId = 0
@@ -19,7 +22,7 @@ const DEFAULT_RESEND_TIMEOUT = 5000
 export default class AbstractSubscription extends EventEmitter {
     constructor(
         streamId, streamPartition, callback, groupKeys,
-        propagationTimeout = DEFAULT_PROPAGATION_TIMEOUT, resendTimeout = DEFAULT_RESEND_TIMEOUT
+        propagationTimeout = DEFAULT_PROPAGATION_TIMEOUT, resendTimeout = DEFAULT_RESEND_TIMEOUT, isCombined = false
     ) {
         super()
         if (!streamId) {
@@ -33,9 +36,30 @@ export default class AbstractSubscription extends EventEmitter {
         this.streamPartition = streamPartition
         this.id = generateSubscriptionId()
         this.groupKeys = groupKeys || {}
-        this.propagationTimeout = propagationTimeout
-        this.resendTimeout = resendTimeout
+
+        if (!isCombined) {
+            this.orderingUtil = new OrderingUtil(streamId, streamPartition, (orderedMessage) => {
+                const newGroupKey = EncryptionUtil.decryptStreamMessage(orderedMessage, this.groupKeys[orderedMessage.getPublisherId()])
+                if (newGroupKey) {
+                    this.groupKeys[orderedMessage.getPublisherId()] = newGroupKey
+                }
+                callback(orderedMessage.getParsedContent(), orderedMessage)
+                if (orderedMessage.isByeMessage()) {
+                    this.emit('done')
+                }
+            }, (from, to, publisherId, msgChainId) => {
+                this.emit('gap', from, to, publisherId, msgChainId)
+            }, propagationTimeout, resendTimeout)
+        }
         this.state = AbstractSubscription.State.unsubscribed
+    }
+
+    _clearGaps() {
+        this.orderingUtil.clearGaps()
+    }
+
+    stop() {
+        this._clearGaps()
     }
 
     getState() {
@@ -46,6 +70,17 @@ export default class AbstractSubscription extends EventEmitter {
         debug(`Subscription: Stream ${this.streamId} state changed ${this.state} => ${state}`)
         this.state = state
         this.emit(state)
+    }
+
+    handleError(err) {
+        /**
+         * If parsing the (expected) message failed, we should still mark it as received. Otherwise the
+         * gap detection will think a message was lost, and re-request the failing message.
+         */
+        if (err instanceof Errors.InvalidJsonError && err.streamMessage) {
+            this.orderingUtil.markMessageExplicitly(err.streamMessage)
+        }
+        this.emit('error', err)
     }
 
     async _catchAndEmitErrors(fn) {
