--- src/Subscription.js
@@ -1,260 +1,0 @@
-import EventEmitter from 'eventemitter3'
-import debugFactory from 'debug'
-import { Errors, Utils } from 'streamr-client-protocol'
-
-import InvalidSignatureError from './errors/InvalidSignatureError'
-import VerificationFailedError from './errors/VerificationFailedError'
-import EncryptionUtil from './EncryptionUtil'
-
-const { OrderingUtil } = Utils
-const debug = debugFactory('StreamrClient::Subscription')
-
-let subId = 0
-function generateSubscriptionId() {
-    const id = subId
-    subId += 1
-    return id.toString()
-}
-
-const DEFAULT_PROPAGATION_TIMEOUT = 5000
-const DEFAULT_RESEND_TIMEOUT = 5000
-
-class Subscription extends EventEmitter {
-    constructor(
-        streamId, streamPartition, callback, options, groupKeys,
-        propagationTimeout = DEFAULT_PROPAGATION_TIMEOUT, resendTimeout = DEFAULT_RESEND_TIMEOUT,
-    ) {
-        super()
-
-        if (!streamId) {
-            throw new Error('No stream id given!')
-        }
-
-        if (!callback) {
-            throw new Error('No callback given!')
-        }
-
-        this.id = generateSubscriptionId()
-        this.streamId = streamId
-        this.streamPartition = streamPartition
-        this.resending = false
-        this.resendOptions = options || {}
-        if (this.resendOptions.from != null && this.resendOptions.last != null) {
-            throw new Error(`Multiple resend options active! Please use only one: ${JSON.stringify(this.resendOptions)}`)
-        }
-
-        if (this.resendOptions.msgChainId != null && typeof this.resendOptions.publisherId === 'undefined') {
-            throw new Error('publisherId must be defined as well if msgChainId is defined.')
-        }
-
-        if (this.resendOptions.from == null && this.resendOptions.to != null) {
-            throw new Error('"from" must be defined as well if "to" is defined.')
-        }
-        this.initialResendDone = Object.keys(this.resendOptions).length === 0
-        this.state = Subscription.State.unsubscribed
-        this.groupKeys = groupKeys || {}
-        this.queue = []
-        this.orderingUtil = new OrderingUtil(streamId, streamPartition, (orderedMessage) => {
-            const newGroupKey = EncryptionUtil.decryptStreamMessage(orderedMessage, this.groupKeys[orderedMessage.getPublisherId()])
-            if (newGroupKey) {
-                this.groupKeys[orderedMessage.getPublisherId()] = newGroupKey
-            }
-            callback(orderedMessage.getParsedContent(), orderedMessage)
-            if (orderedMessage.isByeMessage()) {
-                this.emit('done')
-            }
-        }, (from, to, publisherId, msgChainId) => {
-            this.emit('gap', from, to, publisherId, msgChainId)
-        }, propagationTimeout, resendTimeout)
-
-        /** * Message handlers ** */
-
-        this.on('unsubscribed', () => {
-            this._clearGaps()
-            this.setResending(false)
-        })
-
-        this.on('disconnected', () => {
-            this.setState(Subscription.State.unsubscribed)
-            this._clearGaps()
-            this.setResending(false)
-        })
-
-        this.on('error', () => {
-            this._clearGaps()
-        })
-    }
-
-    _clearGaps() {
-        this.orderingUtil.clearGaps()
-    }
-
-    stop() {
-        this._clearGaps()
-    }
-
-    async _catchAndEmitErrors(fn) {
-        try {
-            return await fn()
-        } catch (err) {
-            console.error(err)
-            this.emit('error', err)
-            // Swallow rejection
-            return Promise.resolve()
-        }
-    }
-
-    // All the handle* methods should:
-    // - return a promise for consistency
-    // - swallow exceptions and emit them as 'error' events
-
-    async handleBroadcastMessage(msg, verifyFn) {
-        return this._catchAndEmitErrors(() => this._handleMessage(msg, verifyFn, false))
-    }
-
-    async handleResentMessage(msg, verifyFn) {
-        return this._catchAndEmitErrors(() => {
-            if (!this.isResending()) {
-                throw new Error(`There is no resend in progress, but received resent message ${msg.serialize()}`)
-            } else {
-                const handleMessagePromise = this._handleMessage(msg, verifyFn, true)
-                this._lastMessageHandlerPromise = handleMessagePromise
-                return handleMessagePromise
-            }
-        })
-    }
-
-    async handleResending(response) {
-        return this._catchAndEmitErrors(() => {
-            if (!this.isResending()) {
-                throw new Error(`There should be no resend in progress, but received ResendResponseResending message ${response.serialize()}`)
-            }
-            this.emit('resending', response)
-        })
-    }
-
-    async handleResent(response) {
-        return this._catchAndEmitErrors(async () => {
-            if (!this.isResending()) {
-                throw new Error(`There should be no resend in progress, but received ResendResponseResent message ${response.serialize()}`)
-            }
-
-            if (!this._lastMessageHandlerPromise) {
-                throw new Error('Attempting to handle ResendResponseResent, but no messages have been received!')
-            }
-
-            // Delay event emission until the last message in the resend has been handled
-            await this._lastMessageHandlerPromise.then(async () => {
-                try {
-                    this.emit('resent', response)
-                } finally {
-                    this._finishResend()
-                }
-            })
-        })
-    }
-
-    async handleNoResend(response) {
-        return this._catchAndEmitErrors(async () => {
-            if (!this.isResending()) {
-                throw new Error(`There should be no resend in progress, but received ResendResponseNoResend message ${response.serialize()}`)
-            }
-            try {
-                this.emit('no_resend', response)
-            } finally {
-                this._finishResend()
-            }
-        })
-    }
-
-    _finishResend() {
-        this._lastMessageHandlerPromise = null
-        this.setResending(false)
-        this.initialResendDone = true
-        this.checkQueue()
-    }
-
-    async _handleMessage(msg, verifyFn, isResend = false) {
-        if (msg.version !== 31) {
-            throw new Error(`Can handle only StreamMessageV31, not version ${msg.version}`)
-        }
-
-        if (msg.prevMsgRef == null) {
-            debug('handleMessage: prevOffset is null, gap detection is impossible! message: %o', msg)
-        }
-
-        // Make sure the verification is successful before proceeding
-        let valid
-        try {
-            valid = await verifyFn()
-        } catch (cause) {
-            throw new VerificationFailedError(msg, cause)
-        }
-
-        if (!valid) {
-            throw new InvalidSignatureError(msg)
-        }
-
-        this.emit('message received')
-        // we queue real-time messages until the initial resend (subscribe with resend options) is completed.
-        if (!this.initialResendDone && !isResend) {
-            this.queue.push(msg)
-        } else {
-            this.orderingUtil.add(msg)
-        }
-    }
-
-    checkQueue() {
-        if (this.queue.length) {
-            debug('Attempting to process %d queued messages for stream %s', this.queue.length, this.streamId)
-
-            const originalQueue = this.queue
-            this.queue = []
-
-            originalQueue.forEach((msg) => this.orderingUtil.add(msg))
-        }
-    }
-
-    hasResendOptions() {
-        return this.resendOptions.from || this.resendOptions.last > 0
-    }
-
-    getState() {
-        return this.state
-    }
-
-    setState(state) {
-        debug(`Subscription: Stream ${this.streamId} state changed ${this.state} => ${state}`)
-        this.state = state
-        this.emit(state)
-    }
-
-    isResending() {
-        return this.resending
-    }
-
-    setResending(resending) {
-        debug(`Subscription: Stream ${this.streamId} resending: ${resending}`)
-        this.resending = resending
-    }
-
-    handleError(err) {
-        /**
-         * If parsing the (expected) message failed, we should still mark it as received. Otherwise the
-         * gap detection will think a message was lost, and re-request the failing message.
-         */
-        if (err instanceof Errors.InvalidJsonError && err.streamMessage) {
-            this.orderingUtil.markMessageExplicitly(err.streamMessage)
-        }
-        this.emit('error', err)
-    }
-}
-
-Subscription.State = {
-    unsubscribed: 'unsubscribed',
-    subscribing: 'subscribing',
-    subscribed: 'subscribed',
-    unsubscribing: 'unsubscribing',
-}
-
-export default Subscription
